Fixes null location issue

+++ app/src/main/java/com/github/adamr22/weatherapp/domain/remote/NetworkRepository.kt
@@ -4,5 +4,5 @@ import com.github.adamr22.weatherapp.domain.Resource
 import com.github.adamr22.weatherapp.domain.models.WeatherInfo
 
 interface NetworkRepository {
-    suspend fun getWeatherData(lat: Long, lon: Long): Resource<WeatherInfo>
+    suspend fun getWeatherData(lat: Double, lon: Double): Resource<WeatherInfo>
 }
